fix(ModalDialog): move types into main modules
