+++ libs/spark/src/Unstable_Modal/index.ts
@@ -1,1 +1,1 @@
-export * from './Unstable_ModalProps';
+export * from './Unstable_Modal';
