+++ libs/spark/src/Unstable_ModalDialogContentText/Unstable_ModalDialogContentText.tsx
@@ -1,11 +1,27 @@
-import React, { forwardRef } from 'react';
-import Unstable_Typography from '../Unstable_Typography';
-import { OverridableComponent } from '../utils';
+import React, { ElementType, forwardRef } from 'react';
+import Unstable_Typography, {
+  Unstable_TypographyTypeMap,
+} from '../Unstable_Typography';
+import { OverridableComponent, OverrideProps } from '../utils';
 import withStyles, { Styles } from '../withStyles';
-import {
-  Unstable_ModalDialogContentTextClassKey,
-  Unstable_ModalDialogContentTextTypeMap,
-} from './Unstable_ModalDialogContentTextProps';
+
+export interface Unstable_ModalDialogContentTextTypeMap<
+  // eslint-disable-next-line @typescript-eslint/ban-types
+  P = {},
+  D extends ElementType = Unstable_TypographyTypeMap['defaultComponent']
+> {
+  props: P & Unstable_TypographyTypeMap['props'];
+  defaultComponent: D;
+  classKey: Unstable_ModalDialogContentTextClassKey;
+}
+
+export type Unstable_ModalDialogContentTextProps<
+  D extends ElementType = Unstable_ModalDialogContentTextTypeMap['defaultComponent'],
+  // eslint-disable-next-line @typescript-eslint/ban-types
+  P = {}
+> = OverrideProps<Unstable_ModalDialogContentTextTypeMap<P, D>, D>;
+
+export type Unstable_ModalDialogContentTextClassKey = 'root';
 
 const styles: Styles<Unstable_ModalDialogContentTextClassKey> = {
   /* Styles applied to the root element. */
