+++ libs/spark/src/Unstable_ModalDialogTitle/Unstable_ModalDialogTitle.tsx
@@ -1,11 +1,27 @@
 import clsx from 'clsx';
-import React, { forwardRef } from 'react';
-import Unstable_Typography from '../Unstable_Typography';
+import React, { forwardRef, HTMLAttributes, ReactNode } from 'react';
+import Unstable_Typography, {
+  Unstable_TypographyProps,
+} from '../Unstable_Typography';
+import { StandardProps } from '../utils';
 import withStyles, { Styles } from '../withStyles';
-import {
-  Unstable_ModalDialogTitleClassKey,
-  Unstable_ModalDialogTitleProps,
-} from './Unstable_ModalDialogTitleProps';
+
+export interface Unstable_ModalDialogTitleProps
+  extends StandardProps<
+    HTMLAttributes<HTMLDivElement>,
+    Unstable_ModalDialogTitleClassKey
+  > {
+  /**
+   * The content of the component.
+   */
+  children?: ReactNode;
+  /**
+   * Props applied to the `Typography` element.
+   */
+  TypographyProps?: Unstable_TypographyProps;
+}
+
+export type Unstable_ModalDialogTitleClassKey = 'root';
 
 const styles: Styles<Unstable_ModalDialogTitleClassKey> = {
   /* Styles applied to the root element. */
